fix(maps): address Engine review feedback
